LPS-56420 Move this to the group local service as it is handling groups it belongs there
